add useTodoControl hook
